requirements.txt
alembic==1.11.1
blinker==1.6.2
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
dominate==2.8.0
email-validator==2.0.0.post2
Flask==2.3.2
Flask-Bootstrap4==4.0.2
flask-bootstrap5==0.1.dev1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-Script==2.0.6
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
greenlet==2.0.2
idna==3.4
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
mysqlclient==2.2.0
SQLAlchemy==2.0.17
typing_extensions==4.7.1
visitor==0.1.3
Werkzeug==2.3.6
WTForms==3.0.1
zipp==3.15.0
